refactor(desktop): extract AI action functions into useAiActions (F5.9)
summarize, draft-reply, touch-up, run-prompt (single + bulk), the
per-panel dismiss / dismiss-all, regenerate-active, and thread-summary
move into useAiActions. AI STATE and the openIdRef/aiCache/mirror refs
stay in App (loadMessage consumes them) and are passed via deps — the
function bodies are byte-identical (behavior preserved by construction).

App.tsx 1,888 -> 1,652; useAiActions 305 (< 500). unit 61 / e2e 39 green
(summarize/prompt/thread-summary specs cover it), tsc + build clean.
