+++ desktop/frontend/src/App.tsx
@@ -1,11 +1,7 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import {
-  applyBulkPromptStream,
-  applyPromptStream,
   backend,
   DEFAULT_KEYMAP,
-  summarizeStream,
-  threadSummaryStream,
   type AccountInfo,
   type KeyMap,
   type Label,
@@ -13,7 +9,6 @@ import {
   type ConfigInfo,
   type MessageDetail,
   type MessageSummary,
-  type Prompt,
   type SavedQuery,
   type ActionPlanResult,
   type AnalyzerRule,
@@ -24,7 +19,6 @@ import {
   cleanSubject,
 } from "./format";
 import { replyInit, forwardInit } from "./compose";
-import { activeAiPanel } from "./aiPanels";
 import ModalsPrimary from "./ModalsPrimary";
 import ModalsSecondary from "./ModalsSecondary";
 import MessageList from "./MessageList";
@@ -40,6 +34,7 @@ import { type AdvFilters, EMPTY_ADV } from "./advancedSearch";
 import { runCommand } from "./commandRunner";
 import { useActionPlan } from "./useActionPlan";
 import { handleKeyDown } from "./keydownHandler";
+import { useAiActions } from "./useAiActions";
 import type { KeydownCtx } from "./keydownCtx";
 import { useBootstrap } from "./useBootstrap";
 import { useZoom } from "./useZoom";
@@ -696,65 +691,16 @@ export default function App() {
     setBulkProgress("");
   }, []);
 
-  const summarize = useCallback(async (id: string, force = false) => {
-    setSummaryForId(id);
-    setSummarizing(true);
-    if (openIdRef.current === id) setSummary("");
-    setError("");
-    try {
-      let acc = "";
-      const final = await summarizeStream(
-        id,
-        (tok) => {
-          acc += tok;
-          // Only paint into the visible panel while this message is still open.
-          if (openIdRef.current === id) setSummary(acc);
-        },
-        force,
-      );
-      updateAiCache(id, { summary: final });
-      if (openIdRef.current === id) setSummary(final);
-    } catch (e) {
-      setError(String(e));
-      if (openIdRef.current === id) setSummary(null);
-    } finally {
-      setSummarizing(false);
-    }
-  }, [updateAiCache]);
-
-  // generateReply asks the AI to draft a reply, then opens the composer with the
-  // draft prefilled so the user can edit before sending.
-  const generateReply = useCallback(
-    async (d: MessageDetail) => {
-      setGeneratingReply(true);
-      setError("");
-      try {
-        const draft = await backend.GenerateReply(d.id);
-        setCompose({ mode: "reply", originalId: d.id, to: d.from, body: draft });
-      } catch (e) {
-        setError(String(e));
-      } finally {
-        setGeneratingReply(false);
-      }
-    },
-    [],
-  );
-
-  // touchUp reformats the open message's body with the AI and shows the cleaned
-  // version in place of the raw text (revertable).
-  const touchUp = useCallback(async (id: string) => {
-    setTouchingUp(true);
-    setError("");
-    try {
-      const t = await backend.TouchUp(id);
-      setTouchUpText(t);
-      updateAiCache(id, { touchUp: t });
-    } catch (e) {
-      setError(String(e));
-    } finally {
-      setTouchingUp(false);
-    }
-  }, [updateAiCache]);
+  const {
+    summarize, generateReply, touchUp, runPrompt, dismissSummary,
+    dismissPrompt, dismissTouchUp, dismissAI, regenerateActive, summarizeThread,
+  } = useAiActions({
+    detail, bulkMode, selected, aiEnabled, showToast, setError,
+    setSummary, setSummarizing, setSummaryForId, setPromptResult, setPromptLabel, setPromptRunning,
+    setPromptForId, setPromptsOpen, setGeneratingReply, setTouchUpText, setTouchingUp, setCompose,
+    setBulkPromptLabel, setBulkPromptText, openIdRef, aiCache, updateAiCache, summaryRef,
+    promptResultRef, touchUpTextRef, promptLabelRef, runningLabelRef, promptPanelRef,
+  });
 
 
   const openStats = useCallback(async () => {
@@ -864,169 +810,6 @@ export default function App() {
     void backend.OpenGmailWeb(id).catch((e) => setError(String(e)));
   }, []);
 
-  const runPrompt = useCallback(
-    async (prompt: Prompt, force = false) => {
-      const bulk = bulkMode && selected.size > 0;
-      if (!bulk && !detail) return;
-      setPromptsOpen(false);
-      setError("");
-
-      // --- bulk (multi-message) prompt: streams into the bulk modal ----------
-      if (bulk) {
-        setPromptForId(null); // bulk isn't tied to the open message's panel
-        setPromptRunning(true);
-        setBulkPromptLabel(`${prompt.name} · ${selected.size} messages`);
-        setBulkPromptText("");
-        try {
-          let acc = "";
-          const final = await applyBulkPromptStream(
-            [...selected],
-            prompt.id,
-            (tok) => {
-              acc += tok;
-              setBulkPromptText(acc);
-            },
-          );
-          setBulkPromptText(final);
-        } catch (e) {
-          setError(String(e));
-          setBulkPromptText(null);
-        } finally {
-          setPromptRunning(false);
-        }
-        return;
-      }
-
-      // --- single-message prompt --------------------------------------------
-      const launchId = detail!.id;
-      // Reuse a result already generated for this (message, prompt) — no new LLM
-      // call, so dismissing then re-running the same prompt is free. force skips
-      // the cache to regenerate (e.g. after editing the prompt).
-      const cached = force
-        ? undefined
-        : aiCache.current.get(launchId)?.promptResults?.[prompt.id];
-      if (cached) {
-        updateAiCache(launchId, { lastPromptId: prompt.id });
-        setPromptForId(launchId);
-        setPromptLabel(cached.label);
-        setPromptResult(cached.text);
-        showToast(`${cached.label} (cached)`);
-        requestAnimationFrame(() =>
-          promptPanelRef.current?.scrollIntoView({ block: "start", behavior: "smooth" }),
-        );
-        return;
-      }
-
-      setPromptForId(launchId);
-      setPromptRunning(true);
-      setPromptLabel(prompt.name);
-      setPromptResult("");
-      runningLabelRef.current[launchId] = prompt.name;
-      try {
-        let acc = "";
-        const final = await applyPromptStream(
-          launchId,
-          prompt.id,
-          (tok) => {
-            acc += tok;
-            // Only paint into the visible panel while this message is still open.
-            if (openIdRef.current === launchId) setPromptResult(acc);
-          },
-          force,
-        );
-        updateAiCache(launchId, {
-          promptResults: {
-            ...(aiCache.current.get(launchId)?.promptResults ?? {}),
-            [prompt.id]: { text: final, label: prompt.name },
-          },
-          lastPromptId: prompt.id,
-        });
-        if (openIdRef.current === launchId) {
-          setPromptResult(final);
-          setPromptLabel(prompt.name);
-        }
-      } catch (e) {
-        setError(String(e));
-        if (openIdRef.current === launchId) setPromptResult(null);
-      } finally {
-        setPromptRunning(false);
-        delete runningLabelRef.current[launchId];
-      }
-    },
-    [detail, bulkMode, selected, showToast, updateAiCache],
-  );
-
-  // Per-panel dismiss: hide the panel and forget just enough of its cache entry
-  // so it stays closed on return. Summary/touch-up drop their cached text; the
-  // prompt keeps its result but clears lastPromptId (so nothing auto-restores).
-  const dismissSummary = useCallback(
-    (id: string | null) => {
-      setSummary(null);
-      if (id) updateAiCache(id, { summary: undefined });
-    },
-    [updateAiCache],
-  );
-  const dismissPrompt = useCallback(
-    (id: string | null) => {
-      setPromptResult(null);
-      if (id) updateAiCache(id, { lastPromptId: undefined });
-    },
-    [updateAiCache],
-  );
-  const dismissTouchUp = useCallback(
-    (id: string | null) => {
-      setTouchUpText(null);
-      if (id) updateAiCache(id, { touchUp: undefined });
-    },
-    [updateAiCache],
-  );
-
-  // dismissAI closes any open AI panel for the current message (summary / prompt /
-  // reformat). Returns whether it dismissed anything (for the layered Escape).
-  const dismissAI = useCallback(() => {
-    const id = openIdRef.current;
-    let any = false;
-    if (summaryRef.current !== null) {
-      dismissSummary(id);
-      any = true;
-    }
-    if (promptResultRef.current !== null) {
-      dismissPrompt(id);
-      any = true;
-    }
-    if (touchUpTextRef.current !== null) {
-      dismissTouchUp(id);
-      any = true;
-    }
-    return any;
-  }, [dismissSummary, dismissPrompt, dismissTouchUp]);
-
-  // regenerateActive re-runs the AI panel currently shown for the open message:
-  // the summary if one is up, otherwise the last prompt (both force a fresh call).
-  const regenerateActive = useCallback(() => {
-    const id = openIdRef.current;
-    if (!id) return;
-    const kind = activeAiPanel({
-      hasSummary: summaryRef.current !== null,
-      hasPrompt: promptResultRef.current !== null,
-      hasTouchUp: touchUpTextRef.current !== null,
-    });
-    if (kind === "prompt") {
-      const pid = aiCache.current.get(id)?.lastPromptId;
-      if (pid != null)
-        void runPrompt(
-          { id: pid, name: promptLabelRef.current, description: "", category: "" },
-          true,
-        );
-      return;
-    }
-    if (kind === "touchup") {
-      void touchUp(id);
-      return;
-    }
-    // summary is shown, or nothing yet → (re)generate the summary.
-    if (aiEnabled) void summarize(id, true);
-  }, [summarize, runPrompt, touchUp, aiEnabled]);
 
   const saveMessage = useCallback(
     (id: string) => {
@@ -1074,25 +857,6 @@ export default function App() {
     [suggestFor, showToast],
   );
 
-  const summarizeThread = useCallback(async () => {
-    if (!detail) return;
-    setSummarizing(true);
-    setSummary("");
-    setError("");
-    try {
-      let acc = "";
-      const final = await threadSummaryStream(detail.threadId, (tok) => {
-        acc += tok;
-        setSummary(acc);
-      });
-      setSummary(final);
-    } catch (e) {
-      setError(String(e));
-      setSummary(null);
-    } finally {
-      setSummarizing(false);
-    }
-  }, [detail]);
 
   const openQueries = useCallback(async () => {
     setQueriesOpen(true);
